+++ CHANGELOG.md
@@ -33,6 +33,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#2092](https://gitlab.com/meltano/meltano/-/issues/2092) Only install plugins related to plugins of the specified type when running `meltano install <plugin_type> --include-related`
 - [#2093](https://gitlab.com/meltano/meltano/-/issues/2093) Print error when transform plugin is installed before dbt
+- [#2089](https://gitlab.com/meltano/meltano/-/issues/2089) Don't add Airflow and dbt-specific files to Meltano project until plugins are added explicitly
+- [#2089](https://gitlab.com/meltano/meltano/-/issues/2089) Don't add docker-compose.yml to Meltano project by default
 
 
 ## 1.35.0 - (2020-06-09)